+++ src/hooks/useSettings/context.ts
@@ -1,0 +1,13 @@
+'use client';
+
+import { createContext } from 'react';
+
+import { SettingsStoreType } from './store';
+
+export interface SettingsContextType {
+  store: SettingsStoreType;
+}
+
+export const SettingsContext = createContext<SettingsContextType | undefined>(
+  undefined
+);