+++ src/model/hooks/usePad.ts
@@ -38,35 +38,35 @@ export const usePad = (padId?: string) => {
     state.context.pads.find((pad) => pad.id === padId)
   );
 
-  const isPadPlayEnabled = useSelector(
-    project,
-    (state) =>
-      (state.context.arePadInteractionsEnabled &&
-        state.context.settings?.isPadPlayEnabled) ??
-      true
-  );
+  // const isPadPlayEnabled = useSelector(
+  //   project,
+  //   (state) =>
+  //     (state.context.arePadInteractionsEnabled &&
+  //       state.context.settings?.isPadPlayEnabled) ??
+  //     true
+  // );
 
-  const isPadSelectSourceDisabled = useSelector(
-    project,
-    (state) => !!state.context.isPadSelectSourceDisabled
-  );
+  // const isPadSelectSourceDisabled = useSelector(
+  //   project,
+  //   (state) => !!state.context.isPadSelectSourceDisabled
+  // );
 
-  const enablePadSelectSource = useCallback(
-    (isEnabled: boolean = true) => {
-      project.send({
-        type: 'setPadSelectSourceDisabled',
-        isDisabled: !isEnabled
-      });
-    },
-    [project]
-  );
+  // const enablePadSelectSource = useCallback(
+  //   (isEnabled: boolean = true) => {
+  //     project.send({
+  //       type: 'setPadSelectSourceDisabled',
+  //       isDisabled: !isEnabled
+  //     });
+  //   },
+  //   [project]
+  // );
 
-  const enablePlayers = useCallback(
-    (isEnabled: boolean = true) => {
-      project.send({ type: 'setPlayersEnabled', isEnabled });
-    },
-    [project]
-  );
+  // const enablePlayers = useCallback(
+  //   (isEnabled: boolean = true) => {
+  //     project.send({ type: 'setPlayersEnabled', isEnabled });
+  //   },
+  //   [project]
+  // );
 
   // const setPadPlayEnabled = useCallback(
   //   (isEnabled: boolean) => {
@@ -181,17 +181,13 @@ export const usePad = (padId?: string) => {
     isPadOneShot,
     chokeGroup,
     playPriority,
-    isPadPlayEnabled,
-    isPadSelectSourceDisabled,
     isPadAssigned,
-    enablePlayers,
     pad,
     selectedPadId,
     setPadIsOneShot,
     setPadIsLooped,
     setPadVolume,
     setPadPlaybackRate,
-    enablePadSelectSource,
     project,
     padLabel,
     setPadLabel,